requirements.txt
langchain
langchain-openai
langchain_community
langchain-core
langchain-astradb
pypdf
python-dotenv
streamlit

-e .